Changes in true output to reflect new Python version.
Some deprecations were noted and now changed regarding the
legend and using 0 in subplots. Finally some changes to
the x and y limits were done, because a white part was left
in the plots. I hope this is due to the version change and
not some internal errors. The dimensions of the arrays were
right, at least.